move to trace: rich_*
