+++ weave/trace/rich_container.py
@@ -3,7 +3,7 @@ from typing import Any, Generic, Iterable, Iterator, Optional, TypeVar
 
 from rich.table import Table
 
-from weave import rich_pydantic_util
+from weave.trace import rich_pydantic_util
 
 T = TypeVar("T")
